-set default ForeColor to white -refactoring -fixed font rendering bug with float locations (TODO: change locations/sizes to PointI/SizeI?)

+++ branches/new_renderer/Controls/Control.cpp
@@ -508,7 +508,7 @@ namespace OSHGui
 			absoluteLocation = location;
 		}
 
-		geometry->SetTranslation(Drawing::Vector(absoluteLocation.X, absoluteLocation.Y, 0.0f));
+		geometry->SetTranslation(Drawing::Vector((int)absoluteLocation.X, (int)absoluteLocation.Y, 0.0f));
 		//TODO: set clipping here
 
 		for (auto &control : internalControls)
